refactor(MigrationDialog): consolidate migration logic into component
Move migration functions (migrateRecordings, migrateTransformations, migrateTransformationRuns, getMigrationCounts) from separate migration.ts module directly into MigrationDialog component. This consolidates all migration-related functionality in one place.

Changes:
- Delete apps/whispering/src/lib/services/db/migration.ts
- Empty migration-test-data.ts (marked for deletion)
- Inline all migration functions and types into MigrationDialog
- Update imports to use inlined types and functions
- Keep all migration logic self-contained within the component
